rmatics/requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-PyMongo==2.1.0
Flask-Redis==0.3.0
Flask-SocketIO==3.0.2
Flask-SQLAlchemy==2.3.2
gevent==1.3.6
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
pymongo==3.7.1
python-engineio==2.2.0
python-socketio==2.0.0
redis==2.10.6
requests==2.19.1
short-url==1.2.2
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
Werkzeug==0.14.1
pymysql
webargs
cent
teamcity-messages
mock
pyhamcrest
flask-testing
uwsgi
redlock-py-1.0.8